Lots of refactoring to convert lambda to use context handler
